+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- More API methods are now can be evaluated in const context.
+- The minimum supported Rust version has been increased to 1.89.0.
 
 ## [3.3.0] - 2025-05-05
 ### Added